chore(misunderstood): refactor test files
